+++ CHANGELOG.md
@@ -6,5 +6,5 @@
   * Renamed funnyplaces-error to factual-error
   * Removed support for CrossRef
   * Removed demo.clj and added real unit tests
-
+  * Upgraded google-api-client to 1.8.0-beta